requirements.txt
# streamlit==1.22.0
# requests==2.26.0
# beautifulsoup4==4.10.0
# pandas==1.3.5
# openpyxl==3.0.9
# xlsxwriter==3.0.3
olefile==0.46
protobuf==3.20.0
numpy>=1.22.0
yfinance
Flask
Flask-Cors
streamlit
pandas
Pillow
xlsxwriter
beautifulsoup4
requests
streamlit
selenium
webdriver-manager
playwright
openpyxl